requirements.txt
albucore==0.0.23
albumentations==1.4.24
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
einops==0.8.1
entmax==1.3
fastapi==0.120.1
filelock==3.20.0
flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
hiredis==3.3.0
httptools==0.7.1
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
mpmath==1.3.0
munch==4.0.0
narwhals==2.10.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pix2tex==0.1.4
protobuf==6.33.0
pyarrow==22.0.0
pydantic==2.12.3
pydantic-core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
redis==7.0.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rpds-py==0.28.0
safetensors==0.6.2
scipy==1.16.3
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
st-img-pastebutton==0.0.6
starlette==0.49.1
streamlit==1.50.0
stringzilla==4.2.3
sympy==1.14.0
tenacity==9.1.2
timm==0.5.4
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
werkzeug==3.1.3
x-transformers==0.15.0
